requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
extruct==0.8.0
html-text==0.5.1
html5lib==1.0.1
idna==2.8
isodate==0.6.0
lxml==4.4.2
mf2py==1.1.2
newsapi-python==0.2.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
Unidecode==1.1.1
urllib3==1.25.7
w3lib==1.21.0
webencodings==0.5.1
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2==2.7.4
psycopg2-binary==2.8.3
python-decouple==3.1
whitenoise==4.1.4